+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 -   Fixed name of locale text JSON schema.
 -   Hide login link on landing page if logged in.
 -   Fixed closing text delimiter localization.
+-   Hide comma separator when localizing names and docs.
 
 ## 0.10.4 2024-07-08